Move ensure_clean_shutdown into rewind module (#1528)
Logically fits there better
